test(formal): tighten allocator rounding bound
- Exhaustively prove two-claim rounding over the uint16 domain
- Preserve full-width uint256 fuzz coverage for production arithmetic
- Keep the asserted conservation and dust bounds unchanged

+++ test/formal/GaugeAllocatorRewards.halmos.t.sol
@@ -40,7 +40,7 @@ contract GaugeAllocatorRewardsHalmosTest is SymTest, Test {
         _assertFundingSplit(received, allocatorShareBps);
     }
 
-    function check_claimRoundingCannotExceedBudget(uint32 budget, uint32 firstWeight, uint32 secondWeight) public pure {
+    function check_claimRoundingCannotExceedBudget(uint16 budget, uint16 firstWeight, uint16 secondWeight) public pure {
         _assertClaimRounding(budget, firstWeight, secondWeight);
     }
